requirements.txt
pytest
typer
pydantic
pydantic-settings
PyYAML
PySide6
Pillow
loguru
mss
pygetwindow
pywin32
pydirectinput
pynput
openai
anthropic
google-genai
instructor
opencv-python
